refactor: remove static lifetimes and random refs
As this was a learning project, there's a lot of weirdness going on here
which looks odd now that I've used Rust a lot more. The return of
`&'static Self` from `new` functions was a holdover from Go usage, but
returning an owned type in Rust makes much more sense.